+++ Engine_Revit_UI/Query/ElementIds.cs
@@ -33,7 +33,7 @@ using BH.Engine.Adapters.Revit;
 using BH.oM.Adapters.Revit.Enums;
 using BH.oM.Adapters.Revit.Interface;
 using BH.oM.Data.Requests;
-
+using BH.Engine.Revit;
 
 namespace BH.UI.Revit.Engine
 {
