+++ Revit_Engine/Revit_Engine.csproj
@@ -139,6 +139,7 @@
     <Compile Include="Modify\AddRefObject.cs" />
     <Compile Include="Modify\AddRefObjects.cs" />
     <Compile Include="Modify\AddMap.cs" />
+    <Compile Include="Modify\Duplicate.cs" />
     <Compile Include="Modify\SetFamilyTypeName.cs" />
     <Compile Include="Modify\SetFamilyName.cs" />
     <Compile Include="Modify\SetDefaultDiscipline.cs" />
